pyproject.toml
[project]
name = "icloud-cleanup-daemon"
version = "1.0.0"
description = "Daemon for cleaning up iCloud sync conflict files (e.g., 'file 2.csv', 'file 3.csv')"
authors = [{ name = "cloud" }]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "watchdog>=4.0.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
icloud-cleanup = "icloud_cleanup.main:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
